--- projects/back-office/src/app/dashboard/pages/reference-data/add-reference-data/add-reference-data.component.html
@@ -1,89 +1,0 @@
-<h1 mat-dialog-title>{{'referenceData.new' | translate}}</h1>
-<div mat-dialog-content>
-    <form [formGroup]="referenceForm">
-        <div>
-            <mat-form-field appearance="outline">
-                <mat-label>{{'global.name' | translate}}</mat-label>
-                <input matInput formControlName="name" type="text" [placeholder]="'global.namePlaceholder' | translate"/>
-                <mat-error *ngIf="name?.errors">
-                    {{'referenceData.warning' | translate}}
-                </mat-error>
-            </mat-form-field>
-    
-            <!-- Type selection -->
-            <mat-form-field appearance="outline">
-                <mat-label>{{ 'global.type' | translate }}</mat-label>
-                <mat-select formControlName="type">
-                    <mat-option value="static">
-                    {{ 'referenceData.static' | translate }}
-                    </mat-option>
-                    <mat-option value="graphql">
-                    {{ 'referenceData.graphql' | translate }}
-                    </mat-option>
-                    <mat-option value="rest">
-                    {{ 'referenceData.rest' | translate }}
-                    </mat-option>
-                </mat-select>
-            </mat-form-field>
-        </div>
-        
-
-        <!-- Specific to GraphQL and Rest type -->
-        <div *ngIf="['rest', 'graphql'].includes(type)">
-            <div>
-                <mat-form-field appearance="outline">
-                    <mat-label>{{ 'referenceData.apiConfiguration' | translate }}</mat-label>
-                    <mat-select #apiConfSelect formControlName="apiConfiguration" (openedChange)="onOpenSelect($event)">
-                        <mat-option *ngFor="let apiConf of apiConfigurations$ | async" [value]="apiConf.id">{{ apiConf.name }}</mat-option>
-                    </mat-select>
-                </mat-form-field>
-                <mat-form-field appearance="outline">
-                    <mat-label>{{'referenceData.queryName' | translate}}</mat-label>
-                    <input matInput formControlName="queryName" type="text" [placeholder]="'global.queryNamePlaceholder' | translate"/>
-                </mat-form-field>
-            </div>
-            <div>
-                <mat-form-field class="chip-list" appearance="outline">
-                    <mat-label>{{ 'referenceData.fields' | translate }}</mat-label>
-                    <mat-chip-list #chipList aria-label="Fields chip list">
-                      <mat-chip
-                        *ngFor="let field of valueFields"
-                        [selectable]="true"
-                        [removable]="true"
-                        (removed)="remove(field)">
-                        {{ field }}
-                        <safe-icon matChipRemove icon="cancel" [size]="18"></safe-icon>
-                      </mat-chip>
-                      <input
-                        #fieldInput
-                        [matChipInputFor]="chipList"
-                        [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-                        (focusout)="add($event)"
-                        (matChipInputTokenEnd)="add($event)"
-                      />
-                    </mat-chip-list>
-                  </mat-form-field>
-                <mat-form-field appearance="outline">
-                    <mat-label>{{'referenceData.path' | translate}}</mat-label>
-                    <input matInput formControlName="path" type="text" [placeholder]="'global.queryNamePlaceholder' | translate"/>
-                </mat-form-field>
-            </div>
-            
-        </div>
-
-        <!-- Value field to be displayed -->
-        <div *ngIf="type">
-            <mat-form-field appearance="outline">
-                <mat-label>{{ 'global.type' | translate }}</mat-label>
-                <mat-select formControlName="valueField">
-                    <mat-option *ngFor="let value of valueFields" [value]="value">{{ value }}</mat-option>
-                </mat-select>
-            </mat-form-field>
-        </div>
-
-    </form>
-</div>
-<div mat-dialog-actions align="end">
-    <button mat-button (click)="onClose()">{{'action.close' | translate}}</button>
-    <button mat-flat-button [mat-dialog-close]="referenceForm.value" cdkFocusInitial color="primary" [disabled]="!referenceForm.valid">{{'action.create' | translate}}</button>
-</div>
