--- projects/back-office/src/app/dashboard/pages/reference-data/add-reference-data/add-reference-data.component.scss
@@ -1,3 +1,0 @@
-mat-form-field {
-  width: 250px;
-}
