+++ projects/back-office/src/app/dashboard/pages/reference-data/reference-data.component.scss
@@ -1,51 +1,21 @@
-.header {
+.csv-input {
+  display: flex;
+  .mat-form-field {
+    flex: 2;
+  }
+  safe-button {
+    flex: 1;
     display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    align-items: top;
-
-    .filters {
-        .top,
-        .advanced {
-            display: flex;
-            flex-direction: row;
-            align-items: top;
-            gap: 16px;
-            flex-wrap: wrap;
-        }
-        mat-form-field {
-            margin-top: -0.25em;
-
-            #custom-search {
-                position: relative;
-                top: -0.15em;
-            }
-        }
-        ::ng-deep .mat-form-field-flex > .mat-form-field-infix {
-            padding: 0.42em 0px !important;
-        }
-        ::ng-deep .mat-form-field-label-wrapper {
-            top: -1.4em;
-        }
-
-        ::ng-deep
-            .mat-form-field-appearance-outline.mat-form-field-can-float.mat-form-field-should-float
-            .mat-form-field-label {
-            transform: translateY(-1.1em) scale(0.75);
-            width: 133.33333%;
-        }
-    }
-
-    .actions {
-        margin-left: auto;
-    }
+    justify-content: center;
+    align-items: center;
+    margin-top: 0.25em;
+    margin-bottom: 0.25em;
+    padding-bottom: 1.34375em;
+  }
 }
 
-.table-column {
-    padding: 0px 4px;
-}
-
-.action-column {
-    border-left: 1px solid rgba(0, 0, 0, 0.12);
+safe-button {
+  &:last-child ::ng-deep button {
+    margin-top: 10px;
+  }
 }
