+++ projects/back-office/src/app/dashboard/pages/reference-data/reference-data.component.ts
@@ -1,299 +1,441 @@
 import {
-  AfterViewInit,
   Component,
+  ElementRef,
   OnDestroy,
   OnInit,
   ViewChild,
 } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
-import { Apollo, QueryRef } from 'apollo-angular';
 import {
-  PermissionsManagement,
-  PermissionType,
-  ReferenceData,
-  SafeAuthService,
-  SafeConfirmModalComponent,
-  SafeSnackBarService,
-} from '@safe/builder';
-import { Subscription } from 'rxjs';
-import {
-  GetReferenceDatasQueryResponse,
-  GET_REFERENCE_DATAS,
-} from '../../../graphql/queries';
-import {
-  AddReferenceDataMutationResponse,
-  ADD_REFERENCE_DATA,
-  DeleteReferenceDataMutationResponse,
-  DELETE_REFERENCE_DATA,
-} from '../../../graphql/mutations';
-import { Router } from '@angular/router';
+  AbstractControl,
+  FormBuilder,
+  FormGroup,
+  Validators,
+} from '@angular/forms';
+import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { AddReferenceDataComponent } from './add-reference-data/add-reference-data.component';
+import {
+  ReferenceData,
+  SafeSnackBarService,
+  referenceType,
+  ApiConfiguration,
+} from '@safe/builder';
+import { Apollo, QueryRef } from 'apollo-angular';
+import { BehaviorSubject, Observable, Subscription } from 'rxjs';
+import {
+  EditReferenceDataMutationResponse,
+  EDIT_REFERENCE_DATA,
+} from '../../../graphql/mutations';
+import {
+  GetApiConfigurationsQueryResponse,
+  GetReferenceDataQueryResponse,
+  GET_API_CONFIGURATIONS_NAMES,
+  GET_REFERENCE_DATA,
+} from '../../../graphql/queries';
+import { COMMA, ENTER, SPACE, TAB } from '@angular/cdk/keycodes';
+import { MatSelect } from '@angular/material/select';
+import { MatChipInputEvent } from '@angular/material/chips';
 
 const ITEMS_PER_PAGE = 10;
+const SEPARATOR_KEYS_CODE = [ENTER, COMMA, TAB, SPACE];
 
 @Component({
   selector: 'app-reference-data',
   templateUrl: './reference-data.component.html',
   styleUrls: ['./reference-data.component.scss'],
 })
-export class ReferenceDataComponent
-  implements OnInit, OnDestroy, AfterViewInit
-{
+export class ReferenceDataComponent implements OnInit, OnDestroy {
   // === DATA ===
   public loading = true;
-  private referenceDatasQuery!: QueryRef<GetReferenceDatasQueryResponse>;
-  displayedColumns = ['name', 'apiConfiguration', 'actions'];
-  dataSource = new MatTableDataSource<ReferenceData>([]);
-  public cachedReferenceDatas: ReferenceData[] = [];
+  public id = '';
+  public referenceData?: ReferenceData;
+  private apolloSubscription?: Subscription;
 
-  // === PERMISSIONS ===
-  canAdd = false;
-  private authSubscription?: Subscription;
+  // === FORM ===
+  public referenceForm: FormGroup = new FormGroup({});
+  public referenceTypeChoices = Object.values(referenceType);
 
-  // === SORTING ===
-  @ViewChild(MatSort) sort?: MatSort;
+  private apiConfigurationsQuery!: QueryRef<GetApiConfigurationsQueryResponse>;
+  private apiConfigurations = new BehaviorSubject<ApiConfiguration[]>([]);
+  public apiConfigurations$!: Observable<ApiConfiguration[]>;
 
-  // === FILTERS ===
-  public searchText = '';
+  public valueFields: string[] = [];
+  readonly separatorKeysCodes: number[] = SEPARATOR_KEYS_CODE;
 
-  public pageInfo = {
-    pageIndex: 0,
-    pageSize: ITEMS_PER_PAGE,
-    length: 0,
+  public csvValue = '';
+  public newData: any = [];
+
+  @ViewChild('formSelect') apiConfSelect?: MatSelect;
+  @ViewChild('fieldInput') fieldInput?: ElementRef<HTMLInputElement>;
+
+  private pageInfo = {
     endCursor: '',
+    hasNextPage: true,
   };
+
+  get name(): AbstractControl | null {
+    return this.referenceForm.get('name');
+  }
+
+  get type(): string {
+    return this.referenceForm.get('type')?.value;
+  }
 
   constructor(
     private apollo: Apollo,
-    public dialog: MatDialog,
+    private route: ActivatedRoute,
     private snackBar: SafeSnackBarService,
-    private authService: SafeAuthService,
     private router: Router,
-    private translate: TranslateService
+    private formBuilder: FormBuilder,
+    private translateService: TranslateService
   ) {}
 
-  /**
-   * Creates the Reference data query, and subscribes to the query changes.
-   */
   ngOnInit(): void {
-    this.referenceDatasQuery =
-      this.apollo.watchQuery<GetReferenceDatasQueryResponse>({
-        query: GET_REFERENCE_DATAS,
-        variables: {
-          first: ITEMS_PER_PAGE,
-        },
-      });
-
-    this.referenceDatasQuery.valueChanges.subscribe((res) => {
-      this.cachedReferenceDatas = res.data.referenceDatas.edges.map(
-        (x) => x.node
-      );
-      this.dataSource.data = this.cachedReferenceDatas.slice(
-        this.pageInfo.pageSize * this.pageInfo.pageIndex,
-        this.pageInfo.pageSize * (this.pageInfo.pageIndex + 1)
-      );
-      this.pageInfo.length = res.data.referenceDatas.totalCount;
-      this.pageInfo.endCursor = res.data.referenceDatas.pageInfo.endCursor;
-      this.loading = res.loading;
-      this.filterPredicate();
-    });
-
-    this.authSubscription = this.authService.user$.subscribe(() => {
-      this.canAdd = this.authService.userHasClaim(
-        PermissionsManagement.getRightFromPath(
-          this.router.url,
-          PermissionType.create
-        )
-      );
-    });
-  }
-
-  /**
-   * Handles page event.
-   *
-   * @param e page event.
-   */
-  onPage(e: any): void {
-    this.pageInfo.pageIndex = e.pageIndex;
-    // Checks if with new page/size more data needs to be fetched
-    if (
-      (e.pageIndex > e.previousPageIndex ||
-        e.pageSize > this.pageInfo.pageSize) &&
-      e.length > this.cachedReferenceDatas.length
-    ) {
-      // Sets the new fetch quantity of data needed as the page size
-      // If the fetch is for a new page the page size is used
-      let neededSize = e.pageSize;
-      // If the fetch is for a new page size, the old page size is substracted from the new one
-      if (e.pageSize > this.pageInfo.pageSize) {
-        neededSize -= this.pageInfo.pageSize;
-      }
-      this.loading = true;
-      this.referenceDatasQuery.fetchMore({
-        variables: {
-          first: neededSize,
-          afterCursor: this.pageInfo.endCursor,
-        },
-        updateQuery: (prev, { fetchMoreResult }) => {
-          if (!fetchMoreResult) {
-            return prev;
-          }
-          return Object.assign({}, prev, {
-            apiConfigurations: {
-              edges: [
-                ...prev.referenceDatas.edges,
-                ...fetchMoreResult.referenceDatas.edges,
-              ],
-              pageInfo: fetchMoreResult.referenceDatas.pageInfo,
-              totalCount: fetchMoreResult.referenceDatas.totalCount,
-            },
-          });
-        },
-      });
-    } else {
-      this.dataSource.data = this.cachedReferenceDatas.slice(
-        e.pageSize * this.pageInfo.pageIndex,
-        e.pageSize * (this.pageInfo.pageIndex + 1)
-      );
-    }
-    this.pageInfo.pageSize = e.pageSize;
-  }
-
-  /**
-   * Frontend filtering.
-   */
-  private filterPredicate(): void {
-    this.dataSource.filterPredicate = (data: any) =>
-      this.searchText.trim().length === 0 ||
-      (this.searchText.trim().length > 0 &&
-        data.name.toLowerCase().includes(this.searchText.trim()));
-  }
-
-  /**
-   * Applies the filter to the data source.
-   *
-   * @param column Column to filter on.
-   * @param event Value of the filter.
-   */
-  applyFilter(column: string, event: any): void {
-    this.searchText = !!event
-      ? event.target.value.trim().toLowerCase()
-      : this.searchText;
-    this.dataSource.filter = '##';
-  }
-
-  /**
-   * Removes all the filters.
-   */
-  clearAllFilters(): void {
-    this.searchText = '';
-    this.applyFilter('', null);
-  }
-
-  /**
-   * Displays the AddReferenceData modal.
-   * Creates a new reference data on closed if result.
-   */
-  onAdd(): void {
-    const dialogRef = this.dialog.open(AddReferenceDataComponent);
-    dialogRef.afterClosed().subscribe((value) => {
-      if (value) {
-        this.apollo
-          .mutate<AddReferenceDataMutationResponse>({
-            mutation: ADD_REFERENCE_DATA,
-            variables: {
-              name: value.name,
-            },
-          })
-          .subscribe(
-            (res) => {
-              if (res.errors) {
-                this.snackBar.openSnackBar(
-                  this.translate.instant('notification.objectNotCreated', {
-                    type: this.translate.instant(
-                      'referenceData.apiConfiguration'
-                    ),
-                    error: res.errors[0].message,
-                  }),
-                  { error: true }
-                );
-              } else {
-                if (res.data) {
-                  this.router.navigate([
-                    '/referencedata',
-                    res.data.addReferenceData.id,
-                  ]);
-                }
-              }
-            },
-            (err) => {
-              this.snackBar.openSnackBar(err.message, { error: true });
-            }
-          );
-      }
-    });
-  }
-
-  /**
-   * Removes a reference data if authorized.
-   *
-   * @param element Reference data to delete.
-   * @param e click event.
-   */
-  onDelete(element: any, e: any): void {
-    e.stopPropagation();
-    const dialogRef = this.dialog.open(SafeConfirmModalComponent, {
-      data: {
-        title: this.translate.instant('APIConf.delete'),
-        content: this.translate.instant('APIConf.deleteDesc', {
-          name: element.name,
-        }),
-        confirmText: this.translate.instant('action.delete'),
-        cancelText: this.translate.instant('action.cancel'),
-        confirmColor: 'warn',
-      },
-    });
-    dialogRef.afterClosed().subscribe((value) => {
-      if (value) {
-        this.apollo
-          .mutate<DeleteReferenceDataMutationResponse>({
-            mutation: DELETE_REFERENCE_DATA,
-            variables: {
-              id: element.id,
-            },
-          })
-          .subscribe((res) => {
-            if (res && !res.errors) {
+    this.id = this.route.snapshot.paramMap.get('id') || '';
+    if (this.id) {
+      this.apolloSubscription = this.apollo
+        .watchQuery<GetReferenceDataQueryResponse>({
+          query: GET_REFERENCE_DATA,
+          variables: {
+            id: this.id,
+          },
+        })
+        .valueChanges.subscribe(
+          (res) => {
+            if (res.data.referenceData) {
+              this.referenceData = res.data.referenceData;
+              this.csvValue =
+                this.referenceData?.data.length > 0
+                  ? this.convertToCSV(this.referenceData?.data)
+                  : '';
+              this.newData =
+                this.referenceData?.data.length > 0
+                  ? this.referenceData?.data
+                  : [];
+              this.referenceForm = this.formBuilder.group({
+                name: [
+                  this.referenceData?.name,
+                  [Validators.required, Validators.pattern('^[A-Za-z-_]+$')],
+                ],
+                type: [this.referenceData?.type, Validators.required],
+                valueField: [
+                  this.referenceData?.valueField,
+                  Validators.required,
+                ],
+                fields: [this.referenceData?.fields, Validators.required],
+                apiConfiguration: [this.referenceData?.apiConfiguration?.id],
+                query: [this.referenceData?.query],
+                path: [this.referenceData?.path],
+                data: [this.referenceData?.data],
+              });
+              this.valueFields = this.referenceForm?.get('fields')?.value;
+              this.loadApiConfigurations(
+                this.referenceForm?.get('type')?.value
+              );
+              // Adapt validators to the type of reference data
+              this.referenceForm.get('type')?.valueChanges.subscribe((type) => {
+                this.loadApiConfigurations(type);
+              });
+              this.loading = res.data.loading;
+            } else {
               this.snackBar.openSnackBar(
-                this.translate.instant('notification.objectDeleted', {
-                  value: this.translate.instant('table.referenceData'),
-                })
+                this.translateService.instant(
+                  'notification.accessNotProvided',
+                  {
+                    type: this.translateService
+                      .instant('notification.term.resource')
+                      .toLowerCase(),
+                    error: '',
+                  }
+                ),
+                { error: true }
               );
-              this.dataSource.data = this.dataSource.data.filter(
-                (x) => x.id !== element.id
-              );
+              this.router.navigate(['/referencedata']);
             }
-          });
-      }
-    });
+          },
+          (err) => {
+            this.snackBar.openSnackBar(err.message, { error: true });
+            this.router.navigate(['/referencedata']);
+          }
+        );
+    }
   }
 
-  /**
-   * Sets the sort in the view.
-   */
-  ngAfterViewInit(): void {
-    this.dataSource.sort = this.sort || null;
-  }
-
-  /**
-   * Removes all subscriptions.
+  /*  Unsubscribe from the apollo subscription if needed
    */
   ngOnDestroy(): void {
-    if (this.authSubscription) {
-      this.authSubscription.unsubscribe();
+    if (this.apolloSubscription) {
+      this.apolloSubscription.unsubscribe();
     }
+  }
+
+  /* Load all Api Configurations
+   */
+  loadApiConfigurations(type?: any): void {
+    if (['graphql', 'rest'].includes(type)) {
+      this.referenceForm
+        .get('apiConfiguration')
+        ?.setValidators(Validators.required);
+      this.referenceForm.get('query')?.setValidators(Validators.required);
+      this.referenceForm.get('fields')?.setValidators(Validators.required);
+
+      this.apiConfigurationsQuery =
+        this.apollo.watchQuery<GetApiConfigurationsQueryResponse>({
+          query: GET_API_CONFIGURATIONS_NAMES,
+          variables: {
+            first: ITEMS_PER_PAGE,
+          },
+        });
+
+      this.apiConfigurations$ = this.apiConfigurations.asObservable();
+      this.apiConfigurationsQuery.valueChanges.subscribe((res) => {
+        this.apiConfigurations.next(
+          res.data.apiConfigurations.edges.map((x) => x.node)
+        );
+        this.pageInfo = res.data.apiConfigurations.pageInfo;
+        this.loading = res.loading;
+      });
+    } else {
+      this.referenceForm.get('apiConfiguration')?.clearValidators();
+      this.referenceForm.get('query')?.clearValidators();
+      this.referenceForm.get('fields')?.clearValidators();
+    }
+    this.referenceForm?.get('apiConfiguration')?.updateValueAndValidity();
+    this.referenceForm?.get('query')?.updateValueAndValidity();
+    this.referenceForm?.get('fields')?.updateValueAndValidity();
+  }
+
+  /*  Edit the permissions layer.
+   */
+  saveAccess(e: any): void {
+    if (this.apolloSubscription) {
+      this.apolloSubscription.unsubscribe();
+    }
+    this.loading = true;
+    this.apollo
+      .mutate<EditReferenceDataMutationResponse>({
+        mutation: EDIT_REFERENCE_DATA,
+        variables: {
+          id: this.id,
+          permissions: e,
+        },
+      })
+      .subscribe((res) => {
+        if (res.data) {
+          this.referenceData = res.data.editReferenceData;
+          this.loading = res.data.loading;
+        }
+      });
+  }
+
+  /*  Edit the Reference data using referenceForm changes
+   */
+  onSave(): void {
+    if (this.apolloSubscription) {
+      this.apolloSubscription.unsubscribe();
+    }
+    this.loading = true;
+    const variables = { id: this.id };
+    Object.assign(
+      variables,
+      this.referenceForm.value.name !== this.referenceData?.name && {
+        name: this.referenceForm.value.name,
+      },
+      this.referenceForm.value.type !== this.referenceData?.type && {
+        type: this.referenceForm.value.type,
+      },
+      this.referenceForm.value.valueField !==
+        this.referenceData?.valueField && {
+        valueField: this.referenceForm.value.valueField,
+      },
+      this.referenceForm.value.fields !== this.referenceData?.fields && {
+        fields: this.referenceForm.value.fields,
+      }
+    );
+    if (['graphql', 'rest'].includes(this.referenceForm.value.type)) {
+      Object.assign(
+        variables,
+        this.referenceForm.value.apiConfiguration !==
+          this.referenceData?.apiConfiguration?.id && {
+          apiConfiguration: this.referenceForm.value.apiConfiguration,
+        },
+        this.referenceForm.value.path !== this.referenceData?.path && {
+          path: this.referenceForm.value.path,
+        },
+        this.referenceForm.value.query !== this.referenceData?.query && {
+          query: this.referenceForm.value.query,
+        }
+      );
+    } else {
+      Object.assign(
+        variables,
+        this.referenceForm.value.data !== this.referenceData?.data && {
+          data: this.referenceForm.value.data,
+        }
+      );
+    }
+    this.apollo
+      .mutate<EditReferenceDataMutationResponse>({
+        mutation: EDIT_REFERENCE_DATA,
+        variables,
+      })
+      .subscribe((res) => {
+        if (res.errors) {
+          this.snackBar.openSnackBar(
+            this.translateService.instant('notification.objectNotUpdated', {
+              type: this.translateService.instant('table.referenceData'),
+              error: res.errors[0].message,
+            }),
+            { error: true }
+          );
+          this.loading = false;
+        } else {
+          this.referenceData = res.data?.editReferenceData;
+          this.referenceForm.markAsPristine();
+          this.loading = res.data?.loading || false;
+        }
+      });
+  }
+
+  /**
+   * Adds scroll listener to select.
+   *
+   * @param e open select event.
+   */
+  onOpenSelect(e: any): void {
+    if (e && this.apiConfSelect) {
+      const panel = this.apiConfSelect.panel.nativeElement;
+      panel.addEventListener('scroll', (event: any) =>
+        this.loadOnScroll(event)
+      );
+    }
+  }
+
+  /**
+   * Fetches more forms on scroll.
+   *
+   * @param e scroll event.
+   */
+  private loadOnScroll(e: any): void {
+    if (
+      e.target.scrollHeight - (e.target.clientHeight + e.target.scrollTop) <
+      50
+    ) {
+      if (!this.loading && this.pageInfo.hasNextPage) {
+        this.loading = true;
+        this.apiConfigurationsQuery.fetchMore({
+          variables: {
+            first: ITEMS_PER_PAGE,
+            afterCursor: this.pageInfo.endCursor,
+          },
+          updateQuery: (prev, { fetchMoreResult }) => {
+            if (!fetchMoreResult) {
+              return prev;
+            }
+            return Object.assign({}, prev, {
+              apiConfigurations: {
+                edges: [
+                  ...prev.apiConfigurations.edges,
+                  ...fetchMoreResult.apiConfigurations.edges,
+                ],
+                pageInfo: fetchMoreResult.apiConfigurations.pageInfo,
+                totalCount: fetchMoreResult.apiConfigurations.totalCount,
+              },
+            });
+          },
+        });
+      }
+    }
+  }
+
+  /**
+   * Add an object to the chip list.
+   *
+   * @param event input event.
+   */
+  add(event: MatChipInputEvent | any): void {
+    // use setTimeout to prevent add input value on focusout
+    setTimeout(
+      () => {
+        const input =
+          event.type === 'focusout'
+            ? this.fieldInput?.nativeElement
+            : event.input;
+        const value =
+          event.type === 'focusout'
+            ? this.fieldInput?.nativeElement.value
+            : event.value;
+
+        // Add the mail
+        if ((value || '').trim()) {
+          // Deep copy needed for the edition
+          const valueFieldsCopy = [...this.valueFields];
+          valueFieldsCopy.push(value.trim());
+          this.valueFields = valueFieldsCopy;
+        }
+        this.referenceForm?.get('fields')?.setValue(this.valueFields);
+        this.referenceForm?.get('fields')?.updateValueAndValidity();
+        // Reset the input value
+        if (input) {
+          input.value = '';
+        }
+      },
+      event.type === 'focusout' ? 500 : 0
+    );
+  }
+
+  /**
+   * Remove an object from the chip list.
+   *
+   * @param field field to remove.
+   */
+  remove(field: string): void {
+    const index = this.valueFields.indexOf(field);
+    if (index >= 0) {
+      // Deep copy needed for the edition
+      const valueFieldsCopy = [...this.valueFields];
+      valueFieldsCopy.splice(index, 1);
+      this.valueFields = valueFieldsCopy;
+    }
+    this.referenceForm?.get('fields')?.setValue(this.valueFields);
+    this.referenceForm?.get('fields')?.updateValueAndValidity();
+  }
+
+  /*  Validate the CSV input and transform it into JSON object
+   */
+  onValidateCSV(): void {
+    const dataTemp = this.referenceForm?.get('data')?.value;
+    this.csvValue = Array.isArray(dataTemp)
+      ? this.convertToCSV(dataTemp)
+      : dataTemp;
+    this.newData = Array.isArray(dataTemp) ? dataTemp : [];
+    if (!Array.isArray(dataTemp)) {
+      const lines = dataTemp.split('\n');
+      const headers = lines[0].split(',');
+      this.valueFields = headers;
+
+      for (let i = 1; i < lines.length; i++) {
+        if (!lines[i]) continue;
+        const obj: any = {};
+        const currentline = lines[i].split(',');
+
+        for (let j = 0; j < headers.length; j++) {
+          obj[headers[j]] = currentline[j];
+        }
+        this.newData.push(obj);
+      }
+      this.referenceForm?.get('data')?.setValue(this.newData);
+      this.referenceForm?.get('data')?.updateValueAndValidity();
+      this.referenceForm?.get('fields')?.setValue(this.valueFields);
+      this.referenceForm?.get('fields')?.updateValueAndValidity();
+    }
+    return this.newData;
+  }
+
+  /*  Convert the JSON into a csv format
+   */
+  convertToCSV(obj: any): string {
+    const array = [Object.keys(obj[0])].concat(obj);
+
+    return array.map((it) => Object.values(it).toString()).join('\n');
   }
 }
