+++ projects/back-office/src/app/dashboard/pages/reference-datas/add-reference-data/add-reference-data.component.ts
@@ -1,0 +1,41 @@
+import { Component, OnInit } from '@angular/core';
+import {
+  AbstractControl,
+  FormBuilder,
+  FormGroup,
+  Validators,
+} from '@angular/forms';
+import { MatDialogRef } from '@angular/material/dialog';
+
+@Component({
+  selector: 'app-add-reference-data',
+  templateUrl: './add-reference-data.component.html',
+  styleUrls: ['./add-reference-data.component.scss'],
+})
+export class AddReferenceDataComponent implements OnInit {
+  // === REACTIVE FORM ===
+  referenceForm: FormGroup = new FormGroup({});
+
+  get name(): AbstractControl | null {
+    return this.referenceForm.get('name');
+  }
+
+  constructor(
+    private formBuilder: FormBuilder,
+    public dialogRef: MatDialogRef<AddReferenceDataComponent>
+  ) {}
+
+  /*  Build the form.
+   */
+  ngOnInit(): void {
+    this.referenceForm = this.formBuilder.group({
+      name: ['', [Validators.required, Validators.pattern('^[A-Za-z-_]+$')]],
+    });
+  }
+
+  /*  Close the modal without sending data.
+   */
+  onClose(): void {
+    this.dialogRef.close();
+  }
+}
