+++ projects/back-office/src/app/dashboard/pages/reference-datas/reference-datas-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ReferenceDatasComponent } from './reference-datas.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ReferenceDatasComponent,
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ReferenceDatasRoutingModule {}
