+++ projects/back-office/src/app/dashboard/pages/reference-datas/reference-datas.component.html
@@ -1,0 +1,97 @@
+<div class="header">
+  <div class="filters">
+    <div class="top">
+      <mat-form-field appearance="outline">
+        <mat-icon matPrefix>search</mat-icon>
+        <input
+          id="custom-search"
+          matInput
+          [ngModel]="searchText"
+          (keyup)="applyFilter('', $event)"
+          type="search"
+          [placeholder]="'filter.search' | translate"
+        />
+      </mat-form-field>
+    </div>
+  </div>
+
+  <div class="actions">
+    <safe-button
+      icon="add"
+      category="secondary"
+      variant="primary"
+      (click)="onAdd()"
+      *ngIf="canAdd"
+    >
+      {{ 'referenceData.add' | translate }}
+    </safe-button>
+  </div>
+</div>
+
+<table *ngIf="!loading" mat-table [dataSource]="dataSource" matSort>
+  <ng-container matColumnDef="name">
+    <th mat-header-cell *matHeaderCellDef>
+      <span class="headerTitle" mat-sort-header>
+        {{ 'global.name' | translate }}
+      </span>
+    </th>
+    <td mat-cell *matCellDef="let element">{{ element.name }}</td>
+  </ng-container>
+
+  <ng-container matColumnDef="apiConfiguration">
+    <th mat-header-cell *matHeaderCellDef>
+      <span class="headerTitle" mat-sort-header>
+        {{ 'referenceData.apiConfiguration' | translate }}
+      </span>
+    </th>
+    <td mat-cell *matCellDef="let element">{{ element.authType }}</td>
+  </ng-container>
+
+
+  <ng-container matColumnDef="actions">
+    <th mat-header-cell *matHeaderCellDef></th>
+    <td mat-cell *matCellDef="let element">
+      <safe-button
+        [isIcon]="true"
+        icon="more_vert"
+        [matMenuTriggerFor]="menu"
+        (click)="$event.stopPropagation()"
+        [disabled]="!(element.canUpdate || element.canDelete)"
+      >
+      </safe-button>
+      <mat-menu #menu="matMenu">
+        <button
+          mat-menu-item
+          (click)="onDelete(element, $event)"
+          *ngIf="element.canDelete"
+        >
+          <mat-icon>delete</mat-icon>
+          {{ 'action.delete' | translate }}
+        </button>
+      </mat-menu>
+    </td>
+  </ng-container>
+
+  <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+  <tr
+    mat-row
+    *matRowDef="let row; columns: displayedColumns"
+    class="clickable"
+    [routerLink]="['/referencedata', row.id]"
+  ></tr>
+</table>
+<safe-skeleton-table
+  *ngIf="loading"
+  [columns]="['global.name', 'referenceData.apiConfiguration']"
+  [actions]="true"
+>
+</safe-skeleton-table>
+<mat-paginator
+  [disabled]="loading"
+  [pageSize]="pageInfo.pageSize"
+  [pageSizeOptions]="[10, 25, 50]"
+  [length]="pageInfo.length"
+  aria-label="Select page of forms"
+  (page)="onPage($event)"
+>
+</mat-paginator>
