+++ projects/back-office/src/app/dashboard/pages/reference-datas/reference-datas.component.scss
@@ -1,0 +1,51 @@
+.header {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    align-items: top;
+
+    .filters {
+        .top,
+        .advanced {
+            display: flex;
+            flex-direction: row;
+            align-items: top;
+            gap: 16px;
+            flex-wrap: wrap;
+        }
+        mat-form-field {
+            margin-top: -0.25em;
+
+            #custom-search {
+                position: relative;
+                top: -0.15em;
+            }
+        }
+        ::ng-deep .mat-form-field-flex > .mat-form-field-infix {
+            padding: 0.42em 0px !important;
+        }
+        ::ng-deep .mat-form-field-label-wrapper {
+            top: -1.4em;
+        }
+
+        ::ng-deep
+            .mat-form-field-appearance-outline.mat-form-field-can-float.mat-form-field-should-float
+            .mat-form-field-label {
+            transform: translateY(-1.1em) scale(0.75);
+            width: 133.33333%;
+        }
+    }
+
+    .actions {
+        margin-left: auto;
+    }
+}
+
+.table-column {
+    padding: 0px 4px;
+}
+
+.action-column {
+    border-left: 1px solid rgba(0, 0, 0, 0.12);
+}
