+++ projects/back-office/src/app/dashboard/pages/reference-datas/reference-datas.module.ts
@@ -1,0 +1,46 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReferenceDatasRoutingModule } from './reference-datas-routing.module';
+import { ReferenceDatasComponent } from './reference-datas.component';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import {
+  FormsModule as AngularFormsModule,
+  ReactiveFormsModule,
+} from '@angular/forms';
+import { MatDialogModule } from '@angular/material/dialog';
+import { SafeButtonModule, SafeSkeletonTableModule } from '@safe/builder';
+import { MatIconModule } from '@angular/material/icon';
+import { MatButtonModule } from '@angular/material/button';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { TranslateModule } from '@ngx-translate/core';
+import { AddReferenceDataComponent } from './add-reference-data/add-reference-data.component';
+
+@NgModule({
+  declarations: [ReferenceDatasComponent, AddReferenceDataComponent],
+  imports: [
+    CommonModule,
+    MatMenuModule,
+    ReferenceDatasRoutingModule,
+    MatProgressSpinnerModule,
+    MatTableModule,
+    MatSortModule,
+    MatFormFieldModule,
+    MatInputModule,
+    AngularFormsModule,
+    ReactiveFormsModule,
+    MatDialogModule,
+    SafeButtonModule,
+    MatIconModule,
+    MatButtonModule,
+    MatPaginatorModule,
+    TranslateModule,
+    SafeSkeletonTableModule,
+  ],
+  exports: [ReferenceDatasComponent],
+})
+export class ReferenceDatasModule {}
