+++ projects/back-office/src/app/graphql/queries.ts
@@ -1084,6 +1084,7 @@ export const GET_REFERENCE_DATAS = gql`
           id
           name
           apiConfiguration {
+            id
             name
           }
           type
@@ -1134,4 +1135,46 @@ export interface GetReferenceDatasQueryResponse {
     };
     totalCount: number;
   };
+}
+
+// === GET REFERENCE DATA ===
+export const GET_REFERENCE_DATA = gql`
+  query GetReferenceData($id: ID!) {
+    referenceData(id: $id) {
+      id
+      name
+      apiConfiguration {
+        id
+        name
+      }
+      type
+      query
+      fields
+      valueField
+      path
+      data
+      permissions {
+        canSee {
+          id
+          title
+        }
+        canUpdate {
+          id
+          title
+        }
+        canDelete {
+          id
+          title
+        }
+      }
+      canSee
+      canUpdate
+      canDelete
+    }
+  }
+`;
+
+export interface GetReferenceDataQueryResponse {
+  loading: boolean;
+  referenceData: ReferenceData;
 }
